refactor: extract session message collator
