+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/injection/FakeSessionModule.kt
@@ -4,6 +4,7 @@ import io.embrace.android.embracesdk.FakeSessionService
 import io.embrace.android.embracesdk.injection.SessionModule
 import io.embrace.android.embracesdk.session.BackgroundActivityService
 import io.embrace.android.embracesdk.session.SessionHandler
+import io.embrace.android.embracesdk.session.SessionMessageCollator
 import io.embrace.android.embracesdk.session.SessionService
 
 internal class FakeSessionModule(
@@ -11,5 +12,8 @@ internal class FakeSessionModule(
     override val sessionService: SessionService = FakeSessionService()
 ) : SessionModule {
     override val sessionHandler: SessionHandler
+        get() = TODO("Not yet implemented")
+
+    override val sessionMessageCollator: SessionMessageCollator
         get() = TODO("Not yet implemented")
 }
