+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/session/SessionModuleImplTest.kt
@@ -46,6 +46,7 @@ internal class SessionModuleImplTest {
         )
         assertNotNull(module.sessionHandler)
         assertNotNull(module.sessionService)
+        assertNotNull(module.sessionMessageCollator)
         assertNull(module.backgroundActivityService)
     }
 
@@ -67,6 +68,7 @@ internal class SessionModuleImplTest {
         )
         assertNotNull(module.sessionHandler)
         assertNotNull(module.sessionService)
+        assertNotNull(module.sessionMessageCollator)
         assertNotNull(module.backgroundActivityService)
     }
